backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
cryptography==41.0.8
pydantic==2.5.0
pydantic-settings==2.1.0
alembic==1.13.1
redis==5.0.1
celery==5.3.4
fastapi-limiter==0.1.5
fastapi-cors==0.0.6